rename router to contacts-router
